+++ src/main/java/org/opentripplanner/transit/raptor/rangeraptor/multicriteria/arrivals/TransferStopArrival.java
@@ -2,6 +2,7 @@ package org.opentripplanner.transit.raptor.rangeraptor.multicriteria.arrivals;
 
 import org.opentripplanner.transit.raptor.api.transit.RaptorTransfer;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTripSchedule;
+import org.opentripplanner.transit.raptor.api.transit.TransitArrival;
 import org.opentripplanner.transit.raptor.api.view.TransferPathView;
 
 /**
@@ -35,5 +36,10 @@ public final class TransferStopArrival<T extends RaptorTripSchedule> extends Abs
     @Override
     public TransferPathView transferPath() {
         return () -> transfer;
+    }
+
+    @Override
+    public TransitArrival<T> mostResentTransitArrival() {
+        return previous().mostResentTransitArrival();
     }
 }