+++ src/main/java/org/opentripplanner/transit/raptor/rangeraptor/standard/StdTransitWorker.java
@@ -6,6 +6,7 @@ import org.opentripplanner.transit.raptor.api.transit.RaptorTransfer;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTripPattern;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTripSchedule;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTripScheduleBoardOrAlightEvent;
+import org.opentripplanner.transit.raptor.api.transit.TransitArrival;
 import org.opentripplanner.transit.raptor.rangeraptor.RoutingStrategy;
 
 
@@ -78,5 +79,10 @@ public final class StdTransitWorker<T extends RaptorTripSchedule> implements Rou
         onTrip = result.getTrip();
         onTripBoardTime = result.getTime();
         onTripBoardStop = stopIndex;
+    }
+
+    @Override
+    public final TransitArrival<T> previousTransit(int boardStopIndex) {
+        return state.previousTransit(boardStopIndex);
     }
 }