+++ src/test/java/org/opentripplanner/transit/raptor/_data/transit/TestRoute.java
@@ -5,6 +5,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import org.opentripplanner.model.base.ToStringBuilder;
+import org.opentripplanner.transit.raptor.api.transit.RaptorGuaranteedTransferProvider;
 import org.opentripplanner.transit.raptor.api.transit.RaptorRoute;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTimeTable;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTripPattern;
@@ -13,6 +14,9 @@ public class TestRoute implements RaptorRoute<TestTripSchedule>, RaptorTimeTable
 
     private final TestTripPattern pattern;
     private final List<TestTripSchedule> schedules = new ArrayList<>();
+    private final TestTransferProvider transfersFrom = new TestTransferProvider();
+    private final TestTransferProvider transfersTo = new TestTransferProvider();
+
 
     private TestRoute(TestTripPattern pattern) {
         this.pattern = pattern;
@@ -46,6 +50,16 @@ public class TestRoute implements RaptorRoute<TestTripSchedule>, RaptorTimeTable
         return pattern;
     }
 
+    @Override
+    public RaptorGuaranteedTransferProvider<TestTripSchedule> getGuaranteedTransfersTo() {
+        return transfersTo;
+    }
+
+    @Override
+    public RaptorGuaranteedTransferProvider<TestTripSchedule> getGuaranteedTransfersFrom() {
+        return transfersFrom;
+    }
+
     public TestRoute withTimetable(TestTripSchedule ... trips) {
         Collections.addAll(schedules, trips);
         return this;
@@ -65,5 +79,32 @@ public class TestRoute implements RaptorRoute<TestTripSchedule>, RaptorTimeTable
                 .addObj("pattern", pattern)
                 .addObj("schedules", schedules)
                 .toString();
+    }
+
+    void addGuaranteedTxFrom(
+            TestTripSchedule fromTrip,
+            int fromTripIndex,
+            int fromStopPos,
+            TestTripSchedule toTrip,
+            int toStopPos
+    ) {
+        int fromTime = fromTrip.arrival(fromStopPos);
+        this.transfersFrom.addGuaranteedTransfers(
+                toTrip, toStopPos, fromTrip, fromTripIndex, fromStopPos, fromTime
+        );
+    }
+
+    void addGuaranteedTxTo(
+            TestTripSchedule fromTrip,
+            int fromStopPos,
+            TestTripSchedule toTrip,
+            int toTripIndex,
+            int toStopPos
+            ) {
+        final int toTime = toTrip.departure(toStopPos);
+        // This is used in the revers search
+        this.transfersTo.addGuaranteedTransfers(
+                fromTrip, fromStopPos, toTrip, toTripIndex, toStopPos, toTime
+        );
     }
 }